requirements.txt
# OCR
paddleocr==2.8.0
paddlepaddle==2.6.2
opencv-python==4.8.1.78
pillow==10.1.0

# PDF обработка
PyMuPDF==1.23.8
pdf2image==1.16.3

# Векторизация
sentence-transformers==2.2.2
chromadb==0.4.18

# LLM
ollama==0.1.6

# Утилиты
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
pydantic==2.5.2
